requirements.txt
chardet==5.2.0
doc2docx==0.2.4
docx==0.2.4
faiss_cpu==1.8.0
fastapi==0.110.2
GitPython==3.1.37
GitPython==3.1.43
numpy==1.26.4
openai==1.23.6
pydantic==1.10.12
python_docx==1.1.0
Requests==2.31.0
torch==2.2.1
tqdm==4.66.2
ujson==5.4.0
